Update utils_plot.py
edit1

+++ utils_plot.py
@@ -29,7 +29,7 @@ def visualize_att(image_path, seq, alphas, idx2word, endseq='<end>', smooth=True
     for t in range(len(words)):
         if t > 50:
             break
-        plt.subplot(np.ceil(len(words) / 5.), 5, t + 1)
+        plt.subplot(int(np.ceil(len(words) / 5)), 5, t + 1)
 
         plt.text(0, 1, '%s' % (words[t]), color='black', backgroundcolor='white', fontsize=12)
         plt.imshow(image)
